A Basic Settings Menu; Fixed Errors
Just a first basic UI for generating and setting sizes.
Also mage a first "Console message" to "test if the algorithm is running. It revealed a logic error. After fixing that and running the test, it showed that the directions are wrong :(
Probably the "opposite" calculation, I need to rethink it.