Fixed some errors and wrong coupling, updated UI
Due to bad naming I got the indices of my test loop wrong (shouldn't code too late). Which looked like the algorithm is to blame, but it was fine - just die representation was wrong :P
Moved some code out from the generator, as it wasn't needed there and caused wrong coupling and even an unnecessary check in some place.
Added a Script to handle the slider vs input conversion and passing the values to the generator.